+++ neulink/src/main/java/com/neucore/neulink/impl/service/resume/DownloadProgressListener.java
@@ -1,5 +1,5 @@
 package com.neucore.neulink.impl.service.resume;
 
 public interface DownloadProgressListener {
-    public void onDownloadSize(int size);
+    public void onDownloadSize(long size);
 }
